Feat/GI-27 Board CRUD 작업을 위한 브랜치 생성
페이지 최고 갯수를 반환해주는 API 추가

+++ src/main/java/javaiscoffee/groomy/ide/board/BoardController.java
@@ -63,7 +63,7 @@ public class BoardController {
 
     @GetMapping("/page-number")
     public ResponseEntity<?> getBoardPageNumber() {
-        int boardPageNumber = boardService.getBoardPageNumber();
+        long boardPageNumber = boardService.getBoardPageNumber();
         return ResponseEntity.ok(boardPageNumber);
     }
 }
